requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
qrcode>=7.4.2
cryptg>=0.4.0
